Keep drop submission types below debt limits
